1.5-sql_generation/pyproject.toml
[project]
name = "sql_generator"
version = "0.1.0"
description = "SQL Query Generator using Groq API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.13.0",
    "loguru>=0.7.3",
    "python-dotenv==1.0.0"
]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "mypy>=1.19.0",
    "ruff>=0.14.8"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py310"
